classical_sampler/imhk-sampler/requirements.txt
numpy==1.23.5
matplotlib==3.7.1
scipy==1.10.1
scikit-learn==1.2.2
sagemath==9.8